# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "IsoGate prototype.",
    "version": "1.0"
  },
  "paths": {
    "/admin": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets a site admin account.",
        "operationId": "GetAdmin",
        "responses": {
          "200": {
            "$ref": "#/responses/AdminResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Creates or updates a site admin account.",
        "operationId": "SaveAdmin",
        "responses": {
          "200": {
            "$ref": "#/responses/AdminResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          },
          "409": {
            "$ref": "#/responses/AdminResponse"
          }
        }
      }
    },
    "/admins": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets all admin accounts for a given Site UUID.",
        "operationId": "GetAdmins",
        "responses": {
          "200": {
            "$ref": "#/responses/AdminsResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
    "/group": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets a subject group by it's UUID.",
        "operationId": "GetGroup",
        "responses": {
          "200": {
            "$ref": "#/responses/GroupResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Creates or updates a subject group.",
        "operationId": "SaveGroup",
        "responses": {
          "200": {
            "$ref": "#/responses/GroupResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          },
          "409": {
            "$ref": "#/responses/GroupResponse"
          }
        }
      }
    },
    "/groups": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets all subject groups for a given Site UUID.",
        "operationId": "GetGroups",
        "responses": {
          "200": {
            "$ref": "#/responses/GroupsResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
    "/license": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Gets a product license by it's UUID.",
        "operationId": "GetLicense",
        "responses": {
          "200": {
            "$ref": "#/responses/LicenseResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Activates a product license for a given product and machine.",
        "operationId": "SaveLicense",
        "responses": {
          "200": {
            "$ref": "#/responses/LicenseResponse"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Deletes a product license by it's UUID.",
        "operationId": "DeleteLicense",
        "responses": {
          "200": {
            "$ref": "#/responses/HttpResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
    "/site": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Retrieves an existing test site.",
        "operationId": "GetSite",
        "responses": {
          "200": {
            "$ref": "#/responses/SiteResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Creates or updates a test site.",
        "operationId": "SaveSite",
        "responses": {
          "200": {
            "$ref": "#/responses/SiteResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          },
          "409": {
            "$ref": "#/responses/SiteResponse"
          }
        }
      }
    },
...
    "/users/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Starts a new authenticated session for a company account.",
        "operationId": "Login",
        "responses": {
          "200": {
            "$ref": "#/responses/SessionTokenResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "summary": "Destroys an existing account session.",
        "operationId": "Logout",
        "responses": {
          "200": {
            "$ref": "#/responses/HttpResponse"
          },
          "401": {
            "$ref": "#/responses/HttpResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Admin": {
      "description": "Admin model",
      "type": "object",
      "properties": {
        "adminActive": {
          "type": "boolean",
          "x-go-name": "AdminActive"
        },
        "adminCreationDate": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "AdminCreationDate"
        },
        "adminEmail": {
          "type": "string",
          "x-go-name": "AdminEmail"
        },
        "adminID": {
          "type": "string",
          "x-go-name": "AdminID"
        },
        "adminLoginName": {
          "type": "string",
          "x-go-name": "AdminLoginName"
        },
        "adminPassword": {
          "type": "string",
          "x-go-name": "AdminPassword"
        },
        "adminPasswordDate": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "AdminPasswordDate"
        },
        "adminSiteUuid": {
          "type": "string",
          "x-go-name": "AdminSiteUUID"
        },
        "adminStatus": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AdminStatus"
        },
        "adminUuid": {
          "type": "string",
          "x-go-name": "AdminUUID"
        },
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "LocalUpdateDateTime"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ServerUpdateDateTime"
        }
      },
      "x-go-package": "isogate/pkg/models"
    },
...
    "Group": {
      "description": "Group model",
      "type": "object",
      "properties": {
        "groupActive": {
          "type": "boolean",
          "x-go-name": "GroupActive"
        },
        "groupId": {
          "type": "string",
          "x-go-name": "GroupID"
        },
        "groupName": {
          "type": "string",
          "x-go-name": "GroupName"
        },
        "groupUuid": {
          "type": "string",
          "x-go-name": "GroupUUID"
        },
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "LocalUpdateDateTime"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ServerUpdateDateTime"
        }
      },
      "x-go-package": "isogate/pkg/models"
    },
...
  },
  "responses": {
    "AdminResponse": {
      "description": "Admin model",
      "headers": {
        "adminActive": {
          "type": "boolean"
        },
        "adminCreationDate": {
          "type": "string",
          "format": "date-time"
        },
        "adminEmail": {
          "type": "string"
        },
        "adminID": {
          "type": "string"
        },
        "adminLoginName": {
          "type": "string"
        },
        "adminPassword": {
          "type": "string"
        },
        "adminPasswordDate": {
          "type": "string",
          "format": "date-time"
        },
        "adminSiteUuid": {
          "type": "string"
        },
        "adminStatus": {
          "type": "integer",
          "format": "int64"
        },
        "adminUuid": {
          "type": "string"
        },
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "AdminsResponse": {
      "description": "Admins is a collection of Admins",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Admin"
        }
      }
    },
...
    "GroupResponse": {
      "description": "Group model",
      "headers": {
        "groupActive": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "groupUuid": {
          "type": "string"
        },
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "GroupsResponse": {
      "description": "Groups is a collection of groups",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Group"
        }
      }
    },
    "HttpResponse": {
      "description": "An ok http response"
    },
    "LicenseResponse": {
      "description": "License model",
      "headers": {
        "activationDate": {
          "type": "string"
        },
        "applicationName": {
          "type": "string"
        },
        "applicationVariation": {
          "type": "string"
        },
        "applicationVersion": {
          "type": "string"
        },
        "licenseId": {
          "type": "string"
        },
        "licenseUuid": {
          "type": "string"
        },
        "machineKey": {
          "type": "string"
        },
        "serialKey": {
          "type": "string"
        }
      }
    },
    "SessionTokenResponse": {
      "description": "SessionToken model",
      "headers": {
        "CreatedDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "ExpiryDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "Token": {
          "type": "string"
        },
        "Username": {
          "type": "string"
        }
      }
    },
    "SiteResponse": {
      "description": "Site model",
      "headers": {
        "localUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "serverUpdateDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "siteID": {
          "type": "string"
        },
        "siteName": {
          "type": "string"
        },
        "siteUUID": {
          "type": "string"
        }
      }
    },
...
  }
}